gave up on default subcommands, added support for gnu flags
As it turns out, default subcommands create an ambiguity of flags.
That wouldn't be too bad if it wasn't for the fact that the help
flag is almost always a global flag, meaning the help flag becomes
ambiguous between the subcommand and the main command. Because of
that, default subcommands aren't actually that great of an idea